Add code to the data.sql file

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE animals (
     date_of_birth DATE NOT NULL,
     escape_attempts INT NOT NULL,
     neutered BOOLEAN NOT NULL,
-    weight_kg DECIMAL NOT NULL,
+    weight_kg FLOAT NOT NULL,
     PRIMARY KEY(id)
 );
